+++ rasa_core/evaluate.py
@@ -19,12 +19,11 @@ from rasa_core import utils
 from rasa_core.agent import Agent
 from rasa_core.events import ActionExecuted, UserUttered
 from rasa_core.interpreter import NaturalLanguageInterpreter
+from rasa_core.policies import SimplePolicyEnsemble
 from rasa_core.trackers import DialogueStateTracker
 from rasa_core.training.generator import TrainingDataGenerator
 from rasa_core.utils import AvailableEndpoints, pad_list_to_size
-from rasa_nlu.evaluate import (
-    plot_confusion_matrix,
-    get_evaluation_metrics)
+from rasa_nlu.evaluate import plot_confusion_matrix, get_evaluation_metrics
 from rasa_nlu.training_data.formats import MarkdownWriter, MarkdownReader
 
 logger = logging.getLogger(__name__)
@@ -140,19 +139,35 @@ class EvaluationStore(object):
                self.entity_predictions != self.entity_targets or \
                self.action_predictions != self.action_targets
 
-    def serialise_targets(self):
-        targets = self.action_targets + \
-                  self.intent_targets + \
-                  self.entity_targets
+    def serialise_targets(self,
+                          include_actions=True,
+                          include_intents=True,
+                          include_entities=False):
+        targets = []
+        if include_actions:
+            targets += self.action_targets
+        if include_intents:
+            targets += self.intent_targets
+        if include_entities:
+            targets += self.entity_targets
+
         return [json.dumps(t) if isinstance(t, dict) else t for t in targets]
 
-    def serialise_predictions(self):
-        predictions = self.action_predictions + \
-                      self.intent_predictions + \
-                      self.entity_predictions
+    def serialise_predictions(self,
+                              include_actions=True,
+                              include_intents=True,
+                              include_entities=False):
+        predictions = []
 
-        return [json.dumps(p) if isinstance(p, dict) else p
-                for p in predictions]
+        if include_actions:
+            predictions += self.action_predictions
+        if include_intents:
+            predictions += self.intent_predictions
+        if include_entities:
+            predictions += self.entity_predictions
+
+        return [json.dumps(t) if isinstance(t, dict) else t
+                for t in predictions]
 
 
 class WronglyPredictedAction(ActionExecuted):
@@ -163,15 +178,45 @@ class WronglyPredictedAction(ActionExecuted):
 
     type_name = "wrong_action"
 
-    def __init__(self, correct_action, predicted_action, timestamp=None):
-        self.correct_action = correct_action
+    def __init__(self, correct_action, predicted_action,
+                 policy, confidence, timestamp=None):
         self.predicted_action = predicted_action
         super(WronglyPredictedAction, self).__init__(correct_action,
+                                                     policy,
+                                                     confidence,
                                                      timestamp=timestamp)
 
     def as_story_string(self):
-        return "{}   <!-- predicted: {} -->".format(self.correct_action,
+        return "{}   <!-- predicted: {} -->".format(self.action_name,
                                                     self.predicted_action)
+
+
+def _deserialise_entities(entities):
+    if isinstance(entities, str):
+        entities = json.loads(entities)
+
+    return [e for e in entities if isinstance(e, dict)]
+
+
+def _md_format_message(text, intent, entities):
+    message_from_md = MarkdownReader()._parse_training_example(text)
+    deserialised_entities = _deserialise_entities(entities)
+    return MarkdownWriter()._generate_message_md(
+            {"text": message_from_md.text,
+             "intent": intent,
+             "entities": deserialised_entities}
+    )
+
+
+class EndToEndUserUtterance(UserUttered):
+    """End-to-end user utterance.
+
+    Mostly used to print the full end-to-end user message in the
+    `failed_stories.md` output file."""
+
+    def as_story_string(self):
+        message = _md_format_message(self.text, self.intent, self.entities)
+        return "{}: {}".format(self.intent.get("name"), message)
 
 
 class WronglyClassifiedUserUtterance(UserUttered):
@@ -185,42 +230,33 @@ class WronglyClassifiedUserUtterance(UserUttered):
     def __init__(self,
                  text,
                  correct_intent,
-                 predicted_intent,
-                 correct_entities=None,
-                 predicted_entities=None,
-                 timestamp=None):
-        self.text = text
-        self.correct_intent = correct_intent
+                 correct_entities,
+                 parse_data=None,
+                 timestamp=None,
+                 input_channel=None,
+                 predicted_intent=None,
+                 predicted_entities=None):
         self.predicted_intent = predicted_intent
-        self.correct_entities = correct_entities
         self.predicted_entities = predicted_entities
-        super(WronglyClassifiedUserUtterance, self).__init__(
-                text, {"name": self.correct_intent}, timestamp=timestamp)
 
-    def _deserialise_entities(self, entities):
-        if isinstance(entities, str):
-            entities = json.loads(entities)
+        intent = {"name": correct_intent}
 
-        return [e for e in entities if isinstance(e, dict)]
-
-    def _md_format_message(self, text, intent, entities):
-        message_from_md = MarkdownReader()._parse_training_example(text)
-        deserialised_entities = self._deserialise_entities(entities)
-        return MarkdownWriter()._generate_message_md(
-                {"text": message_from_md.text,
-                 "intent": intent,
-                 "entities": deserialised_entities}
-        )
+        super(WronglyClassifiedUserUtterance, self).__init__(text,
+                                                             intent,
+                                                             correct_entities,
+                                                             parse_data,
+                                                             timestamp,
+                                                             input_channel)
 
     def as_story_string(self):
-        correct_message = self._md_format_message(self.text,
-                                                  self.correct_intent,
-                                                  self.correct_entities)
-        predicted_message = self._md_format_message(self.text,
-                                                    self.predicted_intent,
-                                                    self.predicted_entities)
+        correct_message = _md_format_message(self.text,
+                                             self.intent,
+                                             self.entities)
+        predicted_message = _md_format_message(self.text,
+                                               self.predicted_intent,
+                                               self.predicted_entities)
         return ("{}: {}   <!-- predicted: {}: {} -->"
-                "").format(self.correct_intent,
+                "").format(self.intent.get("name"),
                            correct_message,
                            self.predicted_intent,
                            predicted_message)
@@ -274,16 +310,22 @@ def _collect_user_uttered_predictions(event,
     if user_uttered_eval_store.has_prediction_target_mismatch():
         partial_tracker.update(
                 WronglyClassifiedUserUtterance(
-                        event.text, intent_gold, predicted_intent,
-                        user_uttered_eval_store.entity_targets,
-                        user_uttered_eval_store.entity_predictions)
+                        event.text, intent_gold,
+                        user_uttered_eval_store.entity_predictions,
+                        event.parse_data,
+                        event.timestamp,
+                        event.input_channel,
+                        predicted_intent,
+                        user_uttered_eval_store.entity_targets)
         )
         if fail_on_prediction_errors:
             raise ValueError(
                     "NLU model predicted a wrong intent. Failed Story:"
                     " \n\n{}".format(partial_tracker.export_stories()))
     else:
-        partial_tracker.update(event)
+        end_to_end_user_utterance = EndToEndUserUtterance(
+                event.text, event.intent, event.entities)
+        partial_tracker.update(end_to_end_user_utterance)
 
     return user_uttered_eval_store
 
@@ -292,7 +334,7 @@ def _collect_action_executed_predictions(processor, partial_tracker, event,
                                          fail_on_prediction_errors):
     action_executed_eval_store = EvaluationStore()
 
-    action, _, _ = processor.predict_next_action(partial_tracker)
+    action, policy, _ = processor.predict_next_action(partial_tracker)
 
     predicted = action.name()
     gold = event.action_name
@@ -301,7 +343,10 @@ def _collect_action_executed_predictions(processor, partial_tracker, event,
                                             action_targets=gold)
 
     if action_executed_eval_store.has_prediction_target_mismatch():
-        partial_tracker.update(WronglyPredictedAction(gold, predicted))
+        partial_tracker.update(WronglyPredictedAction(gold, predicted,
+                                                      event.policy,
+                                                      event.confidence,
+                                                      event.timestamp))
         if fail_on_prediction_errors:
             raise ValueError(
                     "Model predicted a wrong action. Failed Story: "
@@ -309,13 +354,47 @@ def _collect_action_executed_predictions(processor, partial_tracker, event,
     else:
         partial_tracker.update(event)
 
-    return action_executed_eval_store
+    return action_executed_eval_store, policy
+
+
+def _is_in_training_data(tracker, agent, fail_on_prediction_errors=False):
+    processor = agent.create_processor()
+
+    events = list(tracker.events)
+
+    partial_tracker = DialogueStateTracker.from_events(tracker.sender_id,
+                                                       events[:1],
+                                                       agent.domain.slots)
+
+    in_training_data = True
+    test_in_training_data = True
+
+    for event in events[1:]:
+        if isinstance(event, ActionExecuted):
+            _, policy = \
+                _collect_action_executed_predictions(
+                        processor, partial_tracker, event,
+                        fail_on_prediction_errors
+                )
+            if (test_in_training_data and
+                    policy is not None and
+                    SimplePolicyEnsemble.is_not_memo_policy(policy)):
+                in_training_data = False
+                test_in_training_data = False
+
+    return in_training_data
+
+
+def _in_training_data_fraction(in_training_data_list):
+    try:
+        return sum(in_training_data_list) / len(in_training_data_list)
+    except ZeroDivisionError:
+        return 0
 
 
 def _predict_tracker_actions(tracker, agent, fail_on_prediction_errors=False,
                              use_e2e=False):
     processor = agent.create_processor()
-
     tracker_eval_store = EvaluationStore()
 
     events = list(tracker.events)
@@ -326,7 +405,7 @@ def _predict_tracker_actions(tracker, agent, fail_on_prediction_errors=False,
 
     for event in events[1:]:
         if isinstance(event, ActionExecuted):
-            action_executed_result = \
+            action_executed_result, policy = \
                 _collect_action_executed_predictions(
                         processor, partial_tracker, event,
                         fail_on_prediction_errors
@@ -350,7 +429,7 @@ def collect_story_predictions(
         fail_on_prediction_errors=False,  # type: bool
         use_e2e=False  # type: bool
 ):
-    # type: (...) -> Tuple[EvaluationStore, List[DialogueStateTracker]]
+    # type: (...) -> Tuple[EvaluationStore, List[DialogueStateTracker], float]
     """Test the stories from a file, running them through the stored model."""
 
     story_eval_store = EvaluationStore()
@@ -360,11 +439,17 @@ def collect_story_predictions(
     logger.info("Evaluating {} stories\n"
                 "Progress:".format(len(completed_trackers)))
 
+    is_in_training_data_list = []
+
     for tracker in tqdm(completed_trackers):
         tracker_results, predicted_tracker = \
             _predict_tracker_actions(tracker, agent,
                                      fail_on_prediction_errors, use_e2e)
 
+        is_in_training_data = _is_in_training_data(tracker, agent,
+                                                   fail_on_prediction_errors)
+
+        is_in_training_data_list.append(int(is_in_training_data))
         story_eval_store.merge_store(tracker_results)
 
         if tracker_results.has_prediction_target_mismatch():
@@ -378,12 +463,16 @@ def collect_story_predictions(
     report, precision, f1, accuracy = get_evaluation_metrics(
             [1] * len(completed_trackers), correct_dialogues)
 
+    in_training_data_fraction = \
+        _in_training_data_fraction(is_in_training_data_list)
+
     log_evaluation_table([1] * len(completed_trackers),
                          "END-TO-END" if use_e2e else "CONVERSATION",
                          report, precision, f1, accuracy,
+                         in_training_data_fraction,
                          include_report=False)
 
-    return story_eval_store, failed
+    return story_eval_store, failed, in_training_data_fraction
 
 
 def log_failed_stories(failed, failed_output):
@@ -412,7 +501,7 @@ def run_story_evaluation(resource_name, agent,
     completed_trackers = _generate_trackers(resource_name, agent,
                                             max_stories, use_e2e)
 
-    story_results, failed = collect_story_predictions(
+    story_results, failed, in_training_data = collect_story_predictions(
             completed_trackers, agent, fail_on_prediction_errors, use_e2e)
 
     with warnings.catch_warnings():
@@ -425,30 +514,34 @@ def run_story_evaluation(resource_name, agent,
     if out_file_plot:
         plot_story_evaluation(story_results.action_targets,
                               story_results.action_predictions,
-                              report, precision, f1, accuracy, out_file_plot)
+                              report, precision, f1, accuracy,
+                              in_training_data, out_file_plot)
 
     log_failed_stories(failed, out_file_stories)
 
     return {
-        "story_evaluation": {
-            "report": report,
-            "precision": precision,
-            "f1": f1,
-            "accuracy": accuracy
-        }
+        "report": report,
+        "precision": precision,
+        "f1": f1,
+        "accuracy": accuracy,
+        "in_training_data_fraction": in_training_data,
+        "is_end_to_end_evaluation": use_e2e
     }
 
 
 def log_evaluation_table(golds, name,
                          report, precision, f1, accuracy,
+                         in_training_data_fraction,
                          include_report=True):  # pragma: no cover
     """Log the sklearn evaluation metrics."""
     logger.info("Evaluation Results on {} level:".format(name))
-    logger.info("\tCorrect:   {} / {}".format(int(len(golds) * accuracy),
-                                              len(golds)))
-    logger.info("\tF1-Score:  {:.3f}".format(f1))
-    logger.info("\tPrecision: {:.3f}".format(precision))
-    logger.info("\tAccuracy:  {:.3f}".format(accuracy))
+    logger.info("\tCorrect:          {} / {}"
+                "".format(int(len(golds) * accuracy), len(golds)))
+    logger.info("\tF1-Score:         {:.3f}".format(f1))
+    logger.info("\tPrecision:        {:.3f}".format(precision))
+    logger.info("\tAccuracy:         {:.3f}".format(accuracy))
+    logger.info("\tIn-data fraction: {:.3g}"
+                "".format(in_training_data_fraction))
 
     if include_report:
         logger.info("\tClassification report: \n{}".format(report))
@@ -456,6 +549,7 @@ def log_evaluation_table(golds, name,
 
 def plot_story_evaluation(test_y, predictions,
                           report, precision, f1, accuracy,
+                          in_training_data_fraction,
                           out_file):
     """Plot the results of story evaluation"""
     from sklearn.metrics import confusion_matrix
@@ -464,6 +558,7 @@ def plot_story_evaluation(test_y, predictions,
 
     log_evaluation_table(test_y, "ACTION",
                          report, precision, f1, accuracy,
+                         in_training_data_fraction,
                          include_report=True)
 
     cnf_matrix = confusion_matrix(test_y, predictions)